+++ src/Traits/HasAll.php
@@ -4,7 +4,12 @@ namespace seregazhuk\HeadHunterApi\Traits;
 
 trait HasAll
 {
-    abstract public function getResource($verb = '');
+    /**
+     * @param string $verb
+     * @param array $params
+     * @return mixed
+     */
+    abstract public function getResource($verb = '', array $params = []);
 
     public function all()
     {